refactor: Notification to hooks and Typescript
